Added a vagrant instance that uses fewer resources: 1 CPU instead of 2, 512M RAM instead of 1024. Cleaned up a logic error in the bootstrap script.

+++ bootstrap_vm.sh
@@ -5,6 +5,7 @@ if [ $# -eq 0 ]; then
     exit 1
 fi
 
+CONDALOCK="/home/vagrant/.condainstalled"
 
 if [ "$1" == "32" ]; then
     INSTFILE="Anaconda-2.0.1-Linux-x86.sh"
@@ -42,48 +43,60 @@ else
     sudo -u vagrant cp /vagrant/$INSTFILE /home/vagrant
 fi
 
-# Install Anaconda
-sudo -u vagrant bash /home/vagrant/$INSTFILE -bf -p /home/vagrant/anaconda
+# If Anaconda has been installed, just skip next
+# stanza.
+if [ ! -f "$CONDALOCK" ]; then
 
-# Install some conda packages
-sudo -u vagrant /home/vagrant/anaconda/bin/conda create --yes -n workshop python numpy matplotlib netcdf4 basemap pandas ipython pyzmq jinja2 tornado sympy
+    # Install Anaconda
+    sudo -u vagrant bash /home/vagrant/$INSTFILE -bf -p /home/vagrant/anaconda
 
-# Rich only makes owslib availabe for 64-bit. I'll try to build
-# and make available for 32-bit.
-if [ "$1" == "64" ]; then
-    sudo -u vagrant /home/vagrant/anaconda/bin/conda install --yes --quiet -n workshop -c https://conda.binstar.org/rsignell owslib
-else
-    sudo -u vagrant /home/vagrant/anaconda/bin/conda install --yes --quiet -n workshop -c https://conda.binstar.org/wardf owslib
+    # Install some conda packages
+    sudo -u vagrant /home/vagrant/anaconda/bin/conda create --yes -n workshop python numpy matplotlib netcdf4 basemap pandas ipython pyzmq jinja2 tornado sympy
+
+    # Rich only makes owslib availabe for 64-bit. I'll try to build
+    # and make available for 32-bit.
+    if [ "$1" == "64" ]; then
+        sudo -u vagrant /home/vagrant/anaconda/bin/conda install --yes --quiet -n workshop -c https://conda.binstar.org/rsignell owslib
+    else
+        sudo -u vagrant /home/vagrant/anaconda/bin/conda install --yes --quiet -n workshop -c https://conda.binstar.org/wardf owslib
+    fi
+
+    sudo -u vagrant /home/vagrant/anaconda/bin/conda install --yes --quiet -n workshop -c https://conda.binstar.org/Unidata pyudl
+
 fi
-
-sudo -u vagrant /home/vagrant/anaconda/bin/conda install --yes --quiet -n workshop -c https://conda.binstar.org/Unidata pyudl
 
 #
 # Create a text file explaining how to activate the workshop.
 #
 
 OFILE="/home/vagrant/WORKSHOP_README.txt"
-echo -e "\nUnidata Python Workshop\n" > $OFILE
+
+if [ -f "$OFILE" ]; then
+    rm $OFILE
+fi
+
+echo -e "\n\e[92mUnidata Python Workshop\e[39m\n=======================\n" > $OFILE
 echo -e "This Virtual Machine provides a complete environment" >> $OFILE
 echo -e "for the Unidata Python workshop.\n" >> $OFILE
 echo -e "Usage:\n" >> $OFILE
-echo -e '    $ cd /vagrant' >> $OFILE
-echo -e '    $ ipython notebook' >> $OFILE
+echo -e '    $ cd \e[32m/vagrant\e[39m' >> $OFILE
+echo -e '    $ \e[95mipython notebook\e[39m' >> $OFILE
 echo -e ""
-echo -e "This will launch ipython notebook in firefox over x11." >> $OFILE
-echo -e "Alternatively, you may open a browser on your" >> $OFILE
-echo -e "host machine and navigate to:\n" >> $OFILE
-if [ "$1" == "64" ]; then
-    echo -e "    http://127.0.0.1:8864" >> $OFILE
-else
-    echo -e "    http://127.0.0.1:8832" >> $OFILE
-fi
+echo -e "\nThis will launch ipython notebook in firefox over x11." >> $OFILE
 echo -e "\nIf you want to use an external browser from" >> $OFILE
 echo -e "your host machine, you must start ipython notebook" >> $OFILE
 echo -e "as follows:\n" >> $OFILE
-echo -e "    $ ipython notebook --ip=* --no-browser" >> $OFILE
+echo -e '    $ cd \e[32m/vagrant\e[39m' >> $OFILE
+echo -e "    $ \e[95mipython notebook --ip=* --no-browser\e[39m" >> $OFILE
+echo -e "\nYou may then open a browser on your host machine and" >> $OFILE
+echo -e "navigate to:\n" >> $OFILE
+if [ "$1" == "64" ]; then
+    echo -e "    \e[33mhttp://127.0.0.1:8864\e[39m" >> $OFILE
+else
+    echo -e "    \e[33mhttp://127.0.0.1:8832\e[39m" >> $OFILE
+fi
 
-echo -e "\nNote: the /vagrant/ directory is mapped to" >> $OFILE
+echo -e "\nNote: the \e[32m/vagrant/\e[39m directory is mapped to" >> $OFILE
 echo -e "the unidata-python-workshop/ directory on your" >> $OFILE
 echo -e "host machine.  This way any changes made to the" >> $OFILE
 echo -e "notebooks will not dissapear if the VM is destroyed!\n" >> $OFILE
@@ -93,9 +106,13 @@ echo -e "notebooks will not dissapear if the VM is destroyed!\n" >> $OFILE
 #
 
 # Have the user see the WORKSHOP_README.txt file every time the log in.
-sudo -u vagrant echo 'if [ -f "/home/vagrant/WORKSHOP_README.txt" ]; then' >> /home/vagrant/.bashrc
-sudo -u vagrant echo '  cat /home/vagrant/WORKSHOP_README.txt' >> /home/vagrant/.bashrc
-sudo -u vagrant echo 'fi' >> /home/vagrant/.bashrc
 
+if [ ! -f "$CONDALOCK" ]; then
+
+    sudo -u vagrant echo 'if [ -f "/home/vagrant/WORKSHOP_README.txt" ]; then' >> /home/vagrant/.bashrc
+    sudo -u vagrant echo '  cat /home/vagrant/WORKSHOP_README.txt' >> /home/vagrant/.bashrc
+    sudo -u vagrant echo 'fi' >> /home/vagrant/.bashrc
+    sudo -u vagrant touch "$CONDALOCK"
+fi
 # Clean up
 chown -R vagrant:vagrant /home/vagrant
